Fix migrations path (#71)
* Fix migrations path

When I added a separate db for queues I didn't quite do all of the
necessary work to correctly setup a second database. In particular I
didn't separate the migrations, so when I went to add a new migration it
migrated both databases, which is not what I wanted.

Now the migrations are separate and you can make changes to one db
without affecting the other. This is how it should have been from the
beginning.

* fix migrations_path config

* fix queue schema