Don't roll tasks due today (#117)
Wait until they are overdue until rolling over to the next day.
Including tasks due today made sense when this job was run near the end
of the day, since at that point it was unlikely that the tasks would be
completed anyway. But now that this job runs hourly, we don't want to
roll over tasks until they're actually overdue.

This will fix an issue where rollable tasks were always pushed to to the
next day.