Don't raise if invariable
Instead of always trying to scale down the image dimensions for a
thumbnail, first check if the filetype supports doing so. There are some
videos and stuff that don't support this rendering so it causes an
error, making the whole page inacessible.

With this approach, the file still may not render correctly, since we
still use an image tag to display it, but at least one image not working
won't prevent the whole page from erroring.