Generate variants ahead of time (#213)
Pages display lots of images, so the first time you visit a page, you
can potentially wait quite a while for all of the thumbnail variants to
be generated and stored. To fix that, this generates the thumbnail
variant immediatley when an image is uploaded, so it's ready to go right
away.