Keep input focus when autosearching (#230)
Previously when the search was auto submitted, the resulting response
included both the search form and the results. This lead to turbo
replacing both of those elements in the page, meaning your cursor would
lose focus from the search input.

In reality the only parts of the page that change are the search
results, so now only those are sent back. This way your cursor should
keep focus as you expect.

I also fixed an issue where the respose of the TagSearches#show only
included the tag search form, not the entire page. This meant that
if for some reason turbo didn't work, you'd be redirected to a page with
just the search form and nothing else. Now you should end up on the
correct page regardless of whether you're using turbo or not.