Tag search results are linkable
When searching for a tag to add, you can now click on the tag to open
that tag's show page. That way, if you want to see what other types of
images have the tag, you can do so easily. Previously you had to add the
tag, click that, and then remove the tag.