Fix clicking on a tag search result (#335)
The search result is a link to the tag show page. Because it was
rendered within a turbo frame, however, turbo expects the page result to
also contain the same turbo frame tag. Because it's a totally different
page, however, that turbo frame is not present, and so the link fails.

To fix this, we set `data-turbo=false` so that a full page load is
performed.