Remove backlog
I like the idea of this markdown-native ticket system, but I found
myself frequently wanting to add items as I thought of them throughout
the day, and the friction of having to switch to this repo, start the
backlog service, and then write the ticket was too much. I just wouldn't
add them, and then the idea would be lost.

I'm switching to notion because that's where I track other similar
systems with some success.

Thankfully this is very easy to uninstall.