Upgrade openssl gem (#473)
The ruby openssl gem has been updated to fix [a bug] which was
preventing me from uploading images locally.

This is a default gem in ruby, meaning it's bundled with `ruby` itself.
We can upgrade these to newer versions, but they'll also be upgraded
when the language is upgraded.

I'm adding an entry to this project's `Gemfile` to use the patched
openssl vesion. When ruby is updated to include this patched version, we
can upgrade ruby and remove this note.

[a bug]:
https://github.com/ruby/openssl/issues/949#issuecomment-3370358680